Update workshop adatper for more consistent background task treatment.
Modified background task implementation as part of tornado update
process to support latest versions. The Workshop class now implements,
as examples, _two_ background tasks: one runs as a periodic callback on
the server IOLoop, the other as a sleep loop running in a thread
execution pool. This illustrates the two methods available to run
background tasks without clouding one implemntation with the other.