Updates to odin-data workshop instructions to meet latest odin-data spec.
The workshop instructions for odin-data were previously outdated compared to the latest
tagged versions of the excalibur-detector and odin-data releases. Therefore this update
allows the project to be build with versions 0-10-1 and 1.4.0 respectively.

To this end, the .json configuration is now included in the workshop repository and
copied over to the project install directory, along with the demo packet capture, which
is now curled from aegserver.